% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{electricity}
\alias{electricity}
\title{Sample electricity consumption}
\format{A data frame with 384 observations of 4 variables:
\describe{
\item{interval_start}{datetime at start of the interval, POSIXct}
\item{interval_duration}{duration of the interval (s), double}
\item{consumption_rate}{mean electrical consumption-rate (kW), double}
\item{consumption}{total electrical consumption (kWh), double}
}}
\usage{
electricity
}
\description{
A tidytemps dataset. Completely derived from thin air.
}
\keyword{datasets}
