% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprex.R
\name{mutate_y}
\alias{mutate_y}
\title{reprex for using rlang pronouns in functions}
\usage{
mutate_y(df, const)
}
\arguments{
\item{df}{data frame, has numeric column \code{x}}

\item{const}{numeric, scalar}
}
\value{
data frame: \code{df} with numeric column \code{y}
}
\description{
reprex for using rlang pronouns in functions
}
