% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mute_date.R
\name{mute_date}
\alias{mute_date}
\title{Characterize a data frame using date}
\usage{
mute_date(data, duration_ref = lubridate::dhours(0))
}
\arguments{
\item{data}{data frame, compliant with \code{assert_tidytemps()}}

\item{duration_ref}{numeric or \link[lubridate:Duration-class]{lubridate Duration},
number of seconds after local midnight at the start of the date}
}
\value{
\code{data} with two new columns, \code{date} and \code{date_index}
}
\description{
Used when we want to characterize a tidytemps data frame by date and time
within the date. Please make sure that the timezone is set for \code{interval_start}.
}
\details{
Adds two additional columns: (TODO)
}
