% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{tidytemps_validate}
\alias{tidytemps_validate}
\title{Assert that a data frame conforms to tidytemps}
\usage{
tidytemps_validate(data)
}
\arguments{
\item{data}{data frame}
}
\value{
\code{data}
}
\description{
This function is used to assert that a data frame conforms to
the "standard":
}
\details{
\itemize{
\item it has columns \code{interval_start} and \code{interval_duration}
\item the column \code{interval_start} is POSIXct (although we should think of relaxing that to numeric)
\item the column \code{interval_duration} is numeric
}
}
